Add files via upload
We tried OpenCV to draw contours around each object in an image and then finding out a way to always detect contour around the utility poles and from there determine the shape(width and height) of the pole. But this method turned out to be subjective to variance in the background, distance, rotation, tilt, etc, and hence was not suitable to real-life images we had which of course had a lot of such scenarios.