Add files via upload
HOG is a feature extraction technique performed on images and then passed to Linear SVM to classify the poles vs non-poles and then apply a sliding window to get the sub-regions that have poles and get the coordinates. But this technique did not give good classification accuracy to proceed further and did not align with what we were trying to achieve in the end.